Version 3.29.81 (based on bleeding_edge revision r24099)
Remove a couple of deprecated APIs that moved to Isolate.

ARM: Make stack limit stricter to account for large buffers in MacroAssembler (Chromium issue 405338).

Reland 24052 - Require V8 to be explicitly initialized before an Isolate is created.

Require V8 to be explicitly initialized before an Isolate is created.

Performance and stability improvements on all platforms.

git-svn-id: http://v8.googlecode.com/svn/trunk@24100 ce2b1a6d-e550-0410-aec6-3dcde31c8c00